Genera excel en Diarios por Motivos

+++ mmc.Modelos/ContabilidadModels/DiarioXMotivoVM.cs
@@ -13,7 +13,7 @@ namespace mmc.Modelos.ContabilidadModels
         public string TIPO { get; set; }
         public DateTime FECHA { get; set; }
         public int TOTAL { get; set; }
-        public int LIQUIDACION { get; set; }
+        public long LIQUIDACION { get; set; }
         public string NOMBRE { get; set; }
         public string CLIENTE { get; set; }
         public long PROMESA { get; set; }
